refactor(native): split runtime abi modules and tests by domain
