+++ tests/native/runtime/abi/physics/runtime_physics_world_abi_test.c
@@ -1,4 +1,4 @@
-#include "runtime/abi/physics_abi.h"
+#include "runtime/abi/physics/physics_abi.h"
 #include "runtime/world/world_abi.h"
 
 #include <math.h>
