+++ tests/native/runtime/abi/ui/runtime_ui_abi_test.c
@@ -1,4 +1,4 @@
-#include "runtime/abi/ui_abi.h"
+#include "runtime/abi/ui/ui_abi.h"
 
 #include <stdio.h>
